refactor: move graphql file to separete dir
