+++ libs/frontend/modules/app/src/graphql/App.fragment.graphql
