+++ libs/frontend/modules/app/src/graphql/app.endpoints.graphql
