+++ libs/frontend/modules/app/src/index.ts
@@ -1,8 +1,8 @@
 export * from './provider/AppProvider'
 export * from './provider/AppQueryProvider'
 export * from './store'
+export * from './store/app.endpoints'
 export * from './types/react-redux' // Needed for ts to recognize the type
-export * from './use-cases/app.endpoints'
 export * from './use-cases/create-app'
 export * from './use-cases/delete-app'
 export * from './use-cases/get-apps'
