+++ libs/frontend/modules/app/src/store/index.ts
@@ -1,1 +1,2 @@
+export * from './app.endpoints'
 export * from './appState'
