--- libs/frontend/modules/app/src/use-cases/get-app/GetApp.web.graphql.gen.ts
@@ -1,62 +1,0 @@
-import * as Types from '@codelab/frontend/abstract/codegen'
-
-import { AppFragment } from '../../App.fragment.graphql.gen'
-import { gql } from '@apollo/client'
-import { AppFragmentDoc } from '../../App.fragment.graphql.gen'
-import * as Apollo from '@apollo/client'
-const defaultOptions = {}
-export type GetAppQueryVariables = Types.Exact<{
-  input: Types.GetAppInput
-}>
-
-export type GetAppQuery = { app?: AppFragment | null | undefined }
-
-export const GetAppGql = gql`
-  query GetApp($input: GetAppInput!) {
-    app: getApp(input: $input) {
-      ...App
-    }
-  }
-  ${AppFragmentDoc}
-`
-
-/**
- * __useGetAppQuery__
- *
- * To run a query within a React component, call `useGetAppQuery` and pass it any options that fit your needs.
- * When your component renders, `useGetAppQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useGetAppQuery({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useGetAppQuery(
-  baseOptions: Apollo.QueryHookOptions<GetAppQuery, GetAppQueryVariables>,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useQuery<GetAppQuery, GetAppQueryVariables>(GetAppGql, options)
-}
-export function useGetAppLazyQuery(
-  baseOptions?: Apollo.LazyQueryHookOptions<GetAppQuery, GetAppQueryVariables>,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useLazyQuery<GetAppQuery, GetAppQueryVariables>(
-    GetAppGql,
-    options,
-  )
-}
-export type GetAppQueryHookResult = ReturnType<typeof useGetAppQuery>
-export type GetAppLazyQueryHookResult = ReturnType<typeof useGetAppLazyQuery>
-export type GetAppQueryResult = Apollo.QueryResult<
-  GetAppQuery,
-  GetAppQueryVariables
->
-export function refetchGetAppQuery(variables?: GetAppQueryVariables) {
-  return { query: GetAppGql, variables: variables }
-}
