+++ libs/frontend/modules/app/src/use-cases/get-apps/GetAppsList.tsx
@@ -2,7 +2,7 @@ import { padding, threeGridCol } from '@codelab/frontend/style'
 import { EntityType, useCrudModalForm } from '@codelab/frontend/view/components'
 import { Col, Empty, Row, Spin } from 'antd'
 import React from 'react'
-import { useGetAppsQuery } from '../app.endpoints'
+import { useGetAppsQuery } from '../../store/app.endpoints'
 import { CreateAppButtonNow } from '../create-app'
 import { GetAppsItem } from './GetAppsItem'
